Refactor selections argument to opts
