+++ multi/submitbeaconcommitteeselections.go
@@ -23,13 +23,13 @@ import (
 
 // SubmitBeaconCommitteeSelections submits beacon committee selections.
 func (s *Service) SubmitBeaconCommitteeSelections(ctx context.Context,
-	selections []*apiv1.BeaconCommitteeSelection,
+	opts *api.BeaconCommitteeSelectionOpts,
 ) (
 	*api.Response[[]*apiv1.BeaconCommitteeSelection], error,
 ) {
 	res, err := s.doCall(ctx, func(ctx context.Context, client consensusclient.Service) (any, error) {
 		aggregatedSelections, err := client.(consensusclient.BeaconCommitteeSelectionsSubmitter).
-			SubmitBeaconCommitteeSelections(ctx, selections)
+			SubmitBeaconCommitteeSelections(ctx, opts)
 		if err != nil {
 			return nil, err
 		}
