+++ multi/submitbeaconcommitteeselections_test.go
@@ -18,6 +18,7 @@ import (
 	"testing"
 
 	consensusclient "github.com/attestantio/go-eth2-client"
+	"github.com/attestantio/go-eth2-client/api"
 	apiv1 "github.com/attestantio/go-eth2-client/api/v1"
 	"github.com/attestantio/go-eth2-client/mock"
 	"github.com/attestantio/go-eth2-client/multi"
@@ -51,7 +52,8 @@ func TestSubmitBeaconCommitteeSelections(t *testing.T) {
 	require.NoError(t, err)
 
 	for i := 0; i < 128; i++ {
-		res, err := multiClient.(consensusclient.BeaconCommitteeSelectionsSubmitter).SubmitBeaconCommitteeSelections(ctx, []*apiv1.BeaconCommitteeSelection{})
+		res, err := multiClient.(consensusclient.BeaconCommitteeSelectionsSubmitter).
+			SubmitBeaconCommitteeSelections(ctx, &api.BeaconCommitteeSelectionOpts{Selections: []*apiv1.BeaconCommitteeSelection{}})
 		require.NoError(t, err)
 		require.NotNil(t, res)
 	}
